Add functionality for changing the length of a chord
However, it's done in a really ugly manner. Should be fixed as soon as possible.

+++ ChordPlayer/app/src/main/java/com/pudkipz/chordplayer/util/Chord.java
@@ -69,8 +69,7 @@ public class Chord {
         return denominator;
     }
 
-    public int getLength() {
-        return numerator/denominator;
+    public float getLength() {
+        return (float) numerator/(float) denominator;
     }
-
 }
